requirements.txt
Flask==3.0.2
google-api-python-client==2.123.0
python-dotenv==1.0.1
requests==2.31.0
flask_sqlalchemy
apscheduler
gunicorn
flask_caching
redis
psycopg2-binary
marshmallow
